Move old analytics lib to v2 directory.
